requirements.txt
python-telegram-bot==21.11.1
fastapi==0.128.0
uvicorn==0.40.0
yt-dlp==2025.11.12
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
